% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_maps.R
\docType{data}
\name{biceland}
\alias{biceland}
\title{Icelandic shorelines - mainland and islands}
\format{\code{SpatialPolygonsDataFrame}}
\source{
\url{http://www.lmi.is}
}
\usage{
biceland
}
\description{
The original data from Icelandic National Land Survey (Landmælingar
Íslands - LMÍ).  The original contains 6249 separate Polygons (island and scerries)
with a total of 266502 coordinate points.
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

