% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmstools.R
\name{calc_interval}
\alias{calc_interval}
\title{calc_interval}
\usage{
calc_interval(x, w1 = 0.5, w2 = 0.5)
}
\arguments{
\item{x}{A date vector (class "POSIXct")}

\item{w1}{weight 1}

\item{w2}{weight 2}
}
\value{
a vector
}
\description{
Calculate time difference (minutes) between events
}
\examples{
SI_DATIM <- c(lubridate::ymd_hms("2016-01-01 12:00:00") +
                lubridate::minutes(round(rnorm(12,
                                    seq(60, 720, length.out = 12),
                                    20))))
calc_interval(SI_DATIM)
}

