% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csquare.R
\name{csquare_area}
\alias{csquare_area}
\title{Calculate csquare area}
\usage{
csquare_area(x, method = "geo")
}
\arguments{
\item{x}{A c-square}

\item{method}{Default ("geo") implemented}
}
\description{
Internal function as of now. Think there is a
function in DATRAS that does things better.
}

