% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csquare.R
\name{csquare_lat}
\alias{csquare_lat}
\title{Rounded latitude given csquare resolution}
\usage{
csquare_lat(lat, lon, resolution)
}
\arguments{
\item{lat}{Degrees latitude}

\item{lon}{Degrees longitude}

\item{resolution}{A value specifying the C-square resolution.
Valid values are 10, 5, 1, 0.5, 0.1, 0.05 and 0.01 degree units}
}
\value{
A vector of rounded latitudes
}
\description{
Rounded latitude given csquare resolution
}

