% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_decode.R
\name{decode}
\alias{decode}
\title{Convert rectangle code to lon:lat code convention}
\usage{
decode(x, type)
}
\arguments{
\item{x}{Rectangle code vector}

\item{type}{An atomic character vector containng encoding type. Any of
"ices", "csquare" or "iceland".}
}
\value{
A character vector of lon:lat notational format
}
\description{
Convert rectangle code to lon:lat code convention
}
\note{
Within the piped tidyverse approach the returned character vector
can be split using the \code{tidyr::separate} function.
}

