% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csquare.R
\name{decode_csquare}
\alias{decode_csquare}
\title{Decode C-squares}
\usage{
decode_csquare(x, resolution, baf = 0)
}
\arguments{
\item{x}{A string of csquare codes}

\item{resolution}{A value specifying the returned resolution of C-squares:
10, 5, 1, 0.5, 0.1, 0.05 and 0.01 in degree units}

\item{baf}{a value if default (0) no adjustment made. May only be of use for boundary values (-180/180 and -90/90).}
}
\value{
A data frame with longitude (lon) and latitudes (lat)
}
\description{
C-square: A notation system of c-squares provides a compact
encoding of latitude and longitude coordinates into a machine- and
human-readable code. See https://en.wikipedia.org/wiki/C-squares
}

