% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_decode.R
\name{decode_iceland}
\alias{decode_iceland}
\title{Convert Icelandic rectangles to lon:lat}
\usage{
decode_iceland(x)
}
\arguments{
\item{x}{Icelandic rectangle code, e.g. 522}
}
\description{
Convert Icelandic rectangle notation to rectangle center
coordinates notation

Returns a character vector where the coordinates are separated by ":".
}
\note{
This is a modified version of the \code{geo:r2d} function that
returns a single character vector rather than a \code{data.frame}.

Within the piped tidyverse approach the returned character vector
can be split using the \code{tidyr::separate} function.
}
\author{
Höskuldur Björnsson
}

