% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{degrees}
\alias{degrees}
\title{degrees}
\usage{
degrees(x)
}
\arguments{
\item{x}{A value or text}
}
\description{
Af function that adds the suffix degree to object. Used e.g.
in \code{pretty_coordinates}
}

