% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_maps.R
\docType{data}
\name{depth}
\alias{depth}
\title{Depth contours}
\format{\code{SpatiallinesDataFrame}}
\usage{
depth
}
\description{
Present version is stolen, get base from source
}
\keyword{datasets}

