% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df_2_sldf}
\alias{df_2_sldf}
\title{Converts \code{data.frame} to simple SpatialLinesDataFrame}
\usage{
df_2_sldf(df, col.names = c("lon", "lat", "group"))
}
\arguments{
\item{df}{A data.frame containing lon, lat, group and segment.id}

\item{col.names}{A vector specifying the names of longitude, latitude, group and segment id}
}
\description{
Takes a data frame containing lon, lat, and a grouping id
and returns a Spatial Lines Data Frame.
}
