% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df_2_spdf}
\alias{df_2_spdf}
\title{Converts \code{data.frame} to SpatialPolygonsDataFrame}
\usage{
df_2_spdf(df, col.names = c("lon", "lat", "group", "id"))
}
\arguments{
\item{df}{A data.frame containing lon, lat, group and segment.id}

\item{col.names}{A vector specifying the names of longitude, latitude, group and segment id}
}
\description{
Takes a data frame containing lon, lat, a grouping id (here
depth) and then an id for each segment within the group and returns a
Spatial Polygons Data Frame.
}

