% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_maps.R
\docType{data}
\name{dypi_sjm}
\alias{dypi_sjm}
\title{Depth contours from Siglingamálastofnun Íslans}
\format{\code{SpatiallinesDataFrame}}
\usage{
dypi_sjm
}
\description{
Present version is stolen, get base from source
}
\keyword{datasets}

