% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ices.R
\docType{data}
\name{ecoregion}
\alias{ecoregion}
\title{ecoregion}
\format{\code{data.frame}}
\usage{
ecoregion
}
\description{
A thinned version (\~90\%) of the ICES shapefile

The ecoregions: "Greenland Sea", "Bay of Biscay and the Iberian Coast",
"Azores, "Western Mediterranean Sea", "Ionian Sea and the Central Mediterranean Sea",
"Black Sea", "Adriatic Sea", "Aegean-Levantine Sea", "Celtic Seas", "Baltic Sea",
"Greater North Sea", "Arctic Ocean", "Iceland Sea", "Barents Sea", "Faroes",
"Norwegian Sea", "Oceanic Northeast Atlantic"
}
\author{
einar.hjorleifsson@gmail.com
}
\keyword{datasets}

