% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_log.R
\docType{data}
\name{eez_sldf}
\alias{eez_sldf}
\title{Íslenska efnahagslögsagan ('200 mílur') - Línur}
\format{\preformatted{Formal class 'SpatialLinesDataFrame' [package "sp"] with 4 slots
  ..@ data       :'data.frame':	1 obs. of  3 variables:
  .. ..$ ID  : Factor w/ 1 level "1": 1
  .. ..$ name: Factor w/ 1 level "Iceland": 1
  .. ..$ area: num 763249
  ..@ lines      :List of 1
  .. ..$ :Formal class 'Lines' [package "sp"] with 2 slots
  .. .. .. ..@ Lines:List of 1
  .. .. .. .. ..$ :Formal class 'Line' [package "sp"] with 1 slot
  .. .. .. .. .. .. ..@ coords: num [1:3209, 1:2] -12.2 -12.1 -12 -11.8 -11.7 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : NULL
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "x" "y"
  .. .. .. ..@ ID   : chr "1"
  ..@ bbox       : num [1:2, 1:2] -30.87 59.98 -5.58 69.58
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:2] "x" "y"
  .. .. ..$ : chr [1:2] "min" "max"
  ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
  .. .. ..@ projargs: chr "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"
}}
\usage{
eez_sldf
}
\description{
Íslenska efnahaglögsagan reiknuð út frá grunnlínum
(\code{\link{grunnlina}}) sbr.
\href{http://www.althingi.is/altext/lagas/141b/1979041.html}{Lög um landhelgi, efnahagslögsögu og landgrunn}
Hnit á miðlínu eru fengin úr milliríkjasamningum milli
\href{http://www.althingi.is/altext/133/s/0951.html}{Íslands og Færeyjar} og milli
\href{http://www.althingi.is/altext/122/s/1047.html}{Íslands og Grænlands/Noregs}.
}
\examples{
sp::plot(eez_sldf, col="blue")
sp::plot(landhelgi_sldf,col="red",add=TRUE)
sp::plot(iceland,col="grey90",add=TRUE)
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

