% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_encode.R
\name{encode}
\alias{encode}
\title{Encode coordinates to tile codes}
\usage{
encode(x, y, resolution, type = "zchords")
}
\arguments{
\item{x}{longitudinal decimal degrees}

\item{y}{latitudinal decimal degrees}

\item{resolution}{A numeric vector or length 1 or 2 or an character
vector ("large" or "small"). The meaning and usage depends on type - see
further in Details.}

\item{type}{Encoding type, any of "zchords" (default) or "csquare", "ices" or
"iceland".}
}
\value{
A character vector so same lenght as lon and lat
}
\description{
xxx.
}
\details{
xxx.
}

