% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csquare.R
\name{encode_csquare}
\alias{encode_csquare}
\title{Encode coordinates to C-square}
\usage{
encode_csquare(lon, lat, resolution)
}
\arguments{
\item{lon}{A vector of cecimal degrees longitude}

\item{lat}{A vector or decimal degrees latitude}

\item{resolution}{A value specifying the returned resolution of C-squares:
10, 5, 1, 0.5, 0.1, 0.05 and 0.01 in degree units}
}
\value{
A character vector
}
\description{
C-square: A notation system of c-squares provides a compact
encoding of latitude and longitude coordinates into a machine- and
human-readable code. See https://en.wikipedia.org/wiki/C-squares
}

