% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_encode.R
\name{encode_iceland}
\alias{encode_iceland}
\title{Encode coordinates to Icelandic rectangle}
\usage{
encode_iceland(x, y, resolution)
}
\arguments{
\item{x}{lon}

\item{y}{lat}

\item{resolution}{If missing (default) returns the standard (large)
rectangle. If "sr" returns subrectangle.}
}
\value{
A character vector
}
\description{
Encode coordinates to Icelandic rectangle
}

