% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_encode.R
\name{encode_ices}
\alias{encode_ices}
\title{Encode coordinates to ICES rectangle}
\usage{
encode_ices(x, y, resolution, useI = FALSE)
}
\arguments{
\item{x}{lon}

\item{y}{lat}

\item{resolution}{If missing (default) or "large" returns the standard
ICES rectangle code. If "small" returns ICES subrectangle code.}

\item{useI}{Sometimes used}
}
\value{
A character vector
}
\description{
Encode coordinates to ICES rectangle
}

