% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_encode.R
\name{encode_zchords}
\alias{encode_zchords}
\title{Encode coordinates to zchords}
\usage{
encode_zchords(x, y, dx = 1, dy = 0.5, invalids = TRUE)
}
\arguments{
\item{x}{A vector of decimal degrees longitude.}

\item{y}{A vector or decimal degrees latitude.}

\item{dx}{Rectangle width in decimal degrees.}

\item{dy}{Rectangle height in decimal degrees}

\item{invalids}{A boolean, if TRUE (default) returns NA's if
x outside range of -180 to 180 (longitudes) and if y outside
range of -90 to 90 (latitudes).}
}
\value{
A character vector
}
\description{
The function is primarily aimed at generation of user defined
arbritrary statistical rectangles. The longitudinal and latitudinal resolution
can be specified by the user, e.g. dx = 1 and dy = 0.5 is equivalent to the
ices statistical square resolution. The character code returned is of the form
lon:lat, the numerical specifying the central positions.
}
\details{
The tile encoding system is inspired by the code \code{geo::d2dr}
created by Sigurður Þór Jónsson. In acknowledgment to the author and David,
this tile convention is referred to as zchords (__Z(yggi's)-chords__).
It has the advantages over the ICES statistical rectangle encoding and the
Icelandic native's "Tilkynningaskyldureitur" that it is global. In addition
it has the advantage over the __csquare__ encoding in that the resolution
can be anything and is not limited by having equal decimal longitude and
latitude degree-resolution. The __zchords__ is actually all inclusive,
encompassing all the above encoding system. As an added bonus the code
nomenclature is not chipered.
}

