% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{expand_sp}
\alias{expand_sp}
\title{Expand a spatial polygon by certain seamiles}
\usage{
expand_sp(x, miles, quadsegs = 20)
}
\arguments{
\item{x}{A SpatialPolygon object}

\item{miles}{Distance in miles}

\item{quadsegs}{Number of line segments to use to approximate a quarter circle.}
}
\description{
A simpler wrapper around the \code{rgeos::gBuffer} for objects
whose projection is in a non-coordinate format.
}
\examples{
require(rgdal)
i <- iceland$flaki \%in\% "mainland"
three_miles_from_mainland <- expand_sp(iceland[i,], miles=3)
sp::plot(three_miles_from_mainland, col="red")
sp::plot(iceland, col="grey90", add=TRUE)
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}

