% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{geo_area}
\alias{geo_area}
\title{Calculate area in square kilometers}
\usage{
geo_area(x, group_variable)
}
\arguments{
\item{x}{Spatial object}

\item{group_variable}{character vector specifying variable that separates
polygons}
}
\description{
A simpler wrapper around the \code{rgeos::gArea} for calculating area taking
care of the projection of the object.
}
\examples{
require(rgdal)
geo_area(iceland)
geo_area(eez)
geo_area(skipaflokkur3)
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}

