% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geofunctions.R
\name{geo_convert}
\alias{geo_convert}
\title{Converts degree-minutes-seconds to degree decimals}
\usage{
geo_convert(x, inverse = FALSE)
}
\arguments{
\item{x}{Vector of the form DDMMSS (degree-minutes-seconds)}

\item{inverse}{Which conversion should be undertaken, default from 
degrees-minutes-decimal minutes (DDMMmm) to decimal degrees (DD.dd)}
}
\description{
This is the same function as \code{geo::geoconvert.1} , except function does not
return a printout of "error > 60 min"
}
\author{
Hoskuldur Bjornsson <hoski@hafro.is>
}

