% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geofunctions.R
\name{geo_distance}
\alias{geo_distance}
\title{Geographic distance computation}
\usage{
geo_distance(x1, y1, x2, y2, unit = "nmi")
}
\arguments{
\item{x1}{A vector of longitudes.}

\item{y1}{A vector of latitudes.}

\item{x2}{A vector of longitudes.}

\item{y2}{A vector of latitudes.}

\item{unit}{A atomic character vector, if "nmi" (default) returns
values in nautical miles, any other value returns kilometers}
}
\value{
A numerical vector
}
\description{
dplyr-ized version of geo::arcdist
}

