% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geofunctions.R
\name{geo_inside}
\alias{geo_inside}
\title{Classifies if points are inside a region}
\usage{
geo_inside(x, y, reg)
}
\arguments{
\item{x}{A vector of longitudes.}

\item{y}{A vector of latitudes.}

\item{reg}{A data.frame contain column names lon, lat and Region.}
}
\value{
A vector of TRUE and FALSE
}
\description{
Classifies if points are inside a region
}

