% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geofunctions.R
\name{geo_region}
\alias{geo_region}
\title{Allocate region name to points}
\usage{
geo_region(x, y, reg)
}
\arguments{
\item{x}{A vector of longitudes.}

\item{y}{A vector of latitudes.}

\item{reg}{A data.frame contain column names lon, lat and Region.}
}
\value{
A character vector of region names
}
\description{
Allocate region name to points
}

