% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arcgis.R
\name{get_arcgismap}
\alias{get_arcgismap}
\title{Get an ArcGIS map}
\usage{
get_arcgismap(bbox = c(left = -30, bottom = 62.5, right = -10, top = 67.5),
  r = 1, zoom = 6, maptype = c("World_Imagery", "Ocean_Basemap",
  "NatGeo_World_Map", "World_Physical_Map", "World_Shaded_Relief",
  "World_Street_Map", "World_Terrain_Base", "World_Topo_Map"), crop = TRUE,
  messaging = FALSE, urlonly = FALSE, color = c("color"), force = FALSE,
  where = tempdir(), gg = TRUE, ...)
}
\arguments{
\item{bbox}{a bounding box in the format c(lowerleftlon, lowerleftlat,
upperrightlon, upperrightlat).}

\item{r}{Extention of the bounding box. Default is 1 (Not implemented yet)}

\item{zoom}{a zoom level}

\item{maptype}{a character containing any of the following: "World_Imagery" 
(default), "Ocean_Basemap", "NatGeo_World_Map", "World_Physical_Map",
"World_Shaded_Relief", "World_Street_Map", "World_Terrain_Base" and
"World_Topo_Map".}

\item{crop}{crop raw map tiles to specified bounding box}

\item{messaging}{turn messaging on/off}

\item{urlonly}{return url only}

\item{color}{only "color" implemented}

\item{force}{if the map is on file, should a new map be looked up?}

\item{where}{where should the file drawer be located (without terminating "/")}

\item{gg}{Boolean, if TRUE (default) returns a ggmap object with stripped axes
information}

\item{...}{...}
}
\value{
a ggmap object (a classed raster object with a bounding box attribute)
}
\description{
get_arcgis get_arcgismap accesses an online tile server
(http://server.arcgisonline.com/arcgis/rest/services) and 
downloads/stitches map tiles/formats a map image.
}
\seealso{
\code{\link{ggmap}}
}

