% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sjokort.R
\name{get_sjokort}
\alias{get_sjokort}
\title{Get Sjókort}
\usage{
get_sjokort(bbox = c(left = -30, bottom = 62.5, right = -10, top = 67.5),
  r = 1, zoom = 6, maptype = c("sjm", "norwegian"), crop = TRUE,
  messaging = FALSE, urlonly = FALSE, color = c("color"), force = FALSE,
  where = tempdir(), gg = TRUE, ...)
}
\arguments{
\item{bbox}{a bounding box in the format c(lowerleftlon, lowerleftlat,
upperrightlon, upperrightlat).}

\item{r}{Extention of the bounding box. Default is 1}

\item{zoom}{a zoom level}

\item{maptype}{only one type implemented here "sjm"}

\item{crop}{crop raw map tiles to specified bounding box}

\item{messaging}{turn messaging on/off}

\item{urlonly}{return url only}

\item{color}{only "color" implemented}

\item{force}{if the map is on file, should a new map be looked up?}

\item{where}{where should the file drawer be located (without terminating "/")}

\item{gg}{Boolean, if TRUE (default) returns a ggmap object with stripped axes
information}

\item{...}{...}
}
\value{
a ggmap object (a classed raster object with a bounding box attribute)
}
\description{
\code{get_sjokort} accesses a tile server for an Icelandic Sea chart and
downloads/stitches map tiles/formats a map image.
}
\seealso{
\code{\link{ggmap}}
}

