% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_skyndilokun_data.R
\name{get_skyndilokun_data}
\alias{get_skyndilokun_data}
\title{Ná í upplýsingar um skyndilokun}
\usage{
get_skyndilokun_data(nr, ar)
}
\arguments{
\item{nr}{Númer skyndilokunar}

\item{ar}{Árið. Ef óskilgreint, þá er notað núverandi ár}
}
\description{
XXX
}

