% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisland.R
\docType{package}
\name{gisland}
\alias{gisland}
\alias{gisland-package}
\title{gisland: Miscallanous GIS-function and spatial objects}
\description{
Map data with focus on Icelandic ecoregion but includes also some other
shapes of interest. Includes also some useful functions.
}
\details{
To learn more about, start with the vignettes:
\code{browseVignettes(package = "gisland")}
}

