% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_lmi.R
\name{gl_fs_framkvaemdaleyfi_igildi}
\alias{gl_fs_framkvaemdaleyfi_igildi}
\title{Fiskistofa - get framkvæmdarleyfi}
\usage{
gl_fs_framkvaemdaleyfi_igildi()
}
\value{
An sf object
}
\description{
Fiskistofa - get framkvæmdarleyfi
}
