% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_lmi.R
\name{gl_fs_virk_dragnotaveidisvaedi}
\alias{gl_fs_virk_dragnotaveidisvaedi}
\title{Fiskistofa - get dragnótarsvæði}
\usage{
gl_fs_virk_dragnotaveidisvaedi()
}
\value{
An sf object
}
\description{
Fiskistofa - get dragnótarsvæði
}
