% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_lmi.R
\name{gl_fs_virk_grasleppulokanir}
\alias{gl_fs_virk_grasleppulokanir}
\title{Fiskistofa - get grásleppulokanir}
\usage{
gl_fs_virk_grasleppulokanir()
}
\value{
An sf object
}
\description{
Fiskistofa - get grásleppulokanir
}
