% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_lmi.R
\name{gl_fs_virk_humarveidisvaedi}
\alias{gl_fs_virk_humarveidisvaedi}
\title{Fiskistofa - get humarveiðisvæði}
\usage{
gl_fs_virk_humarveidisvaedi()
}
\value{
An sf object
}
\description{
Fiskistofa - get humarveiðisvæði
}
