% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_lmi.R
\name{gl_fs_virkar_skyndilokanir}
\alias{gl_fs_virkar_skyndilokanir}
\title{Fiskistofa - get skyndilokanir}
\usage{
gl_fs_virkar_skyndilokanir()
}
\value{
An sf object
}
\description{
Fiskistofa - get skyndilokanir
}
