% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_lmi.R
\name{gl_lmi_strandlina}
\alias{gl_lmi_strandlina}
\title{Get strandlínur}
\usage{
gl_lmi_strandlina(make_valid = TRUE)
}
\arguments{
\item{make_valid}{A boolean (default TRUE) to "fix" the shapefile}
}
\value{
An sf object
}
\description{
Get strandlínur
}
