% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{grade}
\alias{grade}
\title{Binning data}
\usage{
grade(x, dx)
}
\arguments{
\item{x}{A numerical vector}

\item{dx}{A value, normally less than or equal to 1}
}
\value{
A numerical vector
}
\description{
Binning data
}
