% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_log.R
\docType{data}
\name{grunnpunktar}
\alias{grunnpunktar}
\title{Grunnpunktar}
\format{\code{SpatialPointsDataFrame}}
\source{
{\url{http://www.althingi.is/altext/lagas/141b/1979041.html}}
}
\usage{
grunnpunktar
}
\description{
Hnit sem eru lögð til grundvallar á útreikningi á landhelgi (12 sjómílur) og
efnhagslögsögu Íslands sbr.
 \href{http://www.althingi.is/altext/lagas/141b/1979041.html}{Lög um landhelgi, efnahagslögsögu og landgrunn}
}
\details{
Athugið að grunnpunkar fyrir Kolbeinseyjar (67°08'9 n.br., 18°41'3 v.lg.),
Hvalbak (64°35'8 n.br., 13°16'6 v.lg.) og ystu annesja og skerja Grímseyjar
eru grunnpunktar stórstraumsfjöruborð.

Sjá einnig \code{\link{landhelgi}}, \code{\link{eez}}.
}
\examples{
grunnpunktar@data
sp::plot(iceland, col="grey90")
sp::plot(grunnpunktar, col="red", add=TRUE)
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

