% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hafnir.R
\docType{data}
\name{hafnir}
\alias{hafnir}
\title{Icelandic harbours}
\format{\code{data.frame}}
\usage{
hafnir
}
\description{
The original data from Oracle afli.stadur, geocoded via ggmap
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

