% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_maps.R
\docType{data}
\name{iceland}
\alias{iceland}
\title{Icelandic shorelines - mainland and islands}
\format{\code{SpatialPolygonsDataFrame}}
\source{
\url{http://www.lmi.is}
}
\usage{
iceland
}
\description{
A thinned version of the original data from Icelandic National Land Survey (Landmælingar
Íslands - LMÍ).  The original contains 6249 separate Polygons (island and scerries)
with a total of 266502 coordinate points. This thinned version contains 508 separate Polygons (islands) with a total of
75086 coordinate points.
}
\details{
The alorithm used results in some x4 reduction in resolution without
much loss in quality, at least in overall fisheries context. The object is though
still x3 times as 'rich' as `geo::bisland` and includes the major islands.
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

