% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ices.R
\docType{data}
\name{icesarea}
\alias{icesarea}
\title{ICES areas}
\format{\code{data.frame}}
\usage{
icesarea
}
\description{
A 90% thinned version of the ICES area shapefile
}
\author{
einar.hjorleifsson@gmail.com
}
\keyword{datasets}

