% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices.R
\name{icesmap}
\alias{icesmap}
\title{Create a layer of ICES map backgrounds}
\usage{
icesmap(base = ecoregion, region, projection = "ortho", fill = "white",
  colour = "grey50", lwd = 0.5, islands.fill = "grey90", ...)
}
\arguments{
\item{base}{A map dataframe containing long, lat and group}

\item{region}{map region}

\item{projection}{projection to use}

\item{fill}{fill colour}

\item{colour}{line colour}

\item{lwd}{line width}

\item{islands.fill}{fill colour for islands}

\item{...}{other argument passed on to geom_polygon}
}
\description{
TO DO: use some nice mapproj from the ggalt package
}

