% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ices.R
\docType{data}
\name{icesrect}
\alias{icesrect}
\title{ICES rectangles}
\format{\code{data.frame}}
\usage{
icesrect
}
\description{
ICES area rectangles as shapefile
}
\author{
einar.hjorleifsson@gmail.com
}
\keyword{datasets}

