% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_log.R
\docType{data}
\name{landhelgi_sldf}
\alias{landhelgi_sldf}
\title{Íslenska landhelgin (12 mílur) - Línur}
\format{\code{SpatialLinesDataFrame}}
\usage{
landhelgi_sldf
}
\description{
Landhelgi Íslands (12 mílur) reiknaðar út frá grunnlínum
(\code{\link{grunnlina}}) sbr.
 \href{http://www.althingi.is/altext/lagas/141b/1979041.html}{Lög um landhelgi, efnahagslögsögu og landgrunn}
}
\examples{
sp::plot(landhelgi_sldf, col="blue")
sp::plot(grunnlina, border="yellow", lwd=2, add=TRUE)
sp::plot(grunnpunktar, col="red", add=TRUE)
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

