% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms.R
\name{lookup_interval_ids}
\alias{lookup_interval_ids}
\title{Lookup an id of an event that fall within an invertal given an additional variable}
\usage{
lookup_interval_ids(x, variable, df, cn = c("start", "end", "variable", "id"))
}
\arguments{
\item{x}{A vector containing an event}

\item{variable}{Additional variable that also resides in df}

\item{df}{a data.frame containing as the first four columns: start time (t1),
end time(t2), the additional variable and id (FT_REF, visir, ...) to be 
returned.}

\item{cn}{The column names in df that contain that contain the start and
end of the breakpoints, the (co)variable and the variable to return.}
}
\value{
a vector
}
\description{
Given a dataframe that contains vectors of start and end (breakpoints) as 
well as an additional (co)variable find the interval
containing each element of x and the (co)variable and return a id 
variable from the dataframe (df)
}
\details{
The inspiration for this function was to match vms pings (x) and vessel id
(variable) to a data frame (trip, fishing event, ...) that contains an
intveral, ie. a start and end time, an additional (co)variable 
(e.g. vessel id) and then some "id" that is returned (fishing trip
id, fishing event id, ...). It may though be of use in other cases that may
not necessarily be related to time.
}

