% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logbooks.R
\name{plot_catch}
\alias{plot_catch}
\title{plot_catch}
\usage{
plot_catch(r, species = 1, brks = c(0.01, 0.1, 1, 10, 50, 500))
}
\arguments{
\item{r}{a raster object}

\item{species}{a numerical denoting species}

\item{brks}{a numerical vector denoting breaks}
}
\description{
Plots the 2013 catch distribution for selected stocks
}

