% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_skyndilokun.R
\name{plot_skyndilokun}
\alias{plot_skyndilokun}
\title{Plotta skyndilokun}
\usage{
plot_skyndilokun(d, r = 5)
}
\arguments{
\item{d}{A dataframe containing lon and lat coordinates}

\item{r}{Extention of the bounding box. Default is 5}
}
\description{
XXX
}

