% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{pretty_coordinates}
\alias{pretty_coordinates}
\title{Pretty coordinates}
\usage{
pretty_coordinates(x, suffix = "")
}
\arguments{
\item{x}{A vector contain data of the format degree-minutes-second (ddmmss)}

\item{suffix}{A character, e.g. N, E, W, S}
}
\description{
Creates pretty coordinates for printing.
}

