% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logbooks.R
\name{raster_logbook}
\alias{raster_logbook}
\title{Creates and plot a raster of effort or catch}
\usage{
raster_logbook(xy, z, Resolution = 1, Levels = c(5, 50, 100, 250, 500, 1000,
  1500), plotit = TRUE, xlim = c(-30, -10), ylim = c(62.25, 67.75))
}
\arguments{
\item{xy}{A dataframe or matrix that contains longitudinal (x) and latitudinal (y)
coordinates.}

\item{z}{A numeric vector containing catch. If missing (default) the value assign
to z is 1, this being equivalent to effort. length of z must be the same as
length of x and y.}

\item{Resolution}{Resolution in km of the raster to be generated}

\item{Levels}{The cut points (binwidths) of the values to display in the plot}

\item{plotit}{A boolean, if TRUE (default) plots a base graph. If FALSE
returns a list containing raster- and ggplot-object}

\item{xlim}{XXX}

\item{ylim}{XXX}
}
\description{
XXX
}

