% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sf_ftp.R
\name{read_as_df_ftp}
\alias{read_as_df_ftp}
\title{Read shapefile from ftp}
\usage{
read_as_df_ftp(dsn, shape_url = "ftp://ftp.hafro.is/pub/reiknid/einar/shapes")
}
\arguments{
\item{dsn}{The name of the shapefile (without extention)}

\item{shape_url}{The ftp address}
}
\value{
a simple feature dataframe
}
\description{
Read shapefile from ftp
}
