% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_hafro.R
\name{read_h2o}
\alias{read_h2o}
\title{read object}
\usage{
read_h2o(typename)

read_h2o(typename)
}
\arguments{
\item{typename}{The typname}
}
\value{
A tibble

A tibble
}
\description{
read object

read object
}
