% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sf_ftp.R
\name{read_sf_ftp}
\alias{read_sf_ftp}
\title{Read shapefile from ftp}
\usage{
read_sf_ftp(name, url = "ftp://ftp.hafro.is/pub/data/shapes")
}
\arguments{
\item{name}{The name of the shapefile}

\item{url}{The ftp directory}
}
\value{
a sf-tibble
}
\description{
Read shapefile from ftp
}
