% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_lmi.R
\name{read_strandlinur}
\alias{read_strandlinur}
\title{Get strandlínur}
\usage{
read_strandlinur(fix = TRUE)
}
\arguments{
\item{fix}{A boolean (default TRUE) to "fix" the shapefile}
}
\value{
An sf object
}
\description{
Get strandlínur
}
