% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reglugerdir.R
\docType{data}
\name{reglugerdir}
\alias{reglugerdir}
\title{Reglugerðir um stjórn fiskveiða}
\format{An object of class \code{SpatialPolygonsDataFrame} with 46 rows and 22 columns.}
\usage{
reglugerdir
}
\description{
Incomplete
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

