% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsquare.R
\name{rsquare_encode}
\alias{rsquare_encode}
\title{An arbritrary rectangle from degrees longitudes and latitudes}
\usage{
rsquare_encode(x, y, dx = 1, dy = 0.5, invalids = TRUE)
}
\arguments{
\item{x}{A vector of decimal degrees longitude.}

\item{y}{A vector or decimal degrees latitude.}

\item{dx, }{dy Rectangle width and height in decimal degrees.}

\item{invalids}{A boolean, if TRUE (default) returns NA's if
x outside range of -180 to 180 (longitudes) and if y oustride
reange of -90 to 90 (latitudes).}
}
\value{
A character vector
}
\description{
Put some readable text here.
}

