% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{scale_latitude}
\alias{scale_latitude}
\title{Labels maps with degree and minutes}
\usage{
scale_latitude(from = -90, to = 90, by = 0.5, letter = FALSE, ...)
}
\arguments{
\item{from}{Minimum latitude label}

\item{to}{Maximum latitude label}

\item{by}{Breaks in decimal degrees}

\item{letter}{Flag which controls if lable should include letters, i.e. N or S. The
is set to FALSE}

\item{...}{other arguments passed to scale_x_continuous}
}
\description{
Labels maps with degree and minutes.
}

