% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{scale_longitude}
\alias{scale_longitude}
\title{Labels maps with degree and minutes}
\usage{
scale_longitude(from = -180, to = 180, by = 1, letter = FALSE, ...)
}
\arguments{
\item{from}{Minimum longitude label}

\item{to}{Maximum longitude label}

\item{by}{Breaks in decimal degrees}

\item{letter}{Flag which controls if lable should include letters, i.e. E or W. The
is set to FALSE}

\item{...}{other arguments passed to scale_x_continuous}
}
\description{
Labels maps with degree and minutes.
}

