% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_log.R
\docType{data}
\name{skipaflokkur1}
\alias{skipaflokkur1}
\title{Heimildir fiskiskipa til veiða í 1. flokki}
\format{\code{SpatialPolygons}}
\usage{
skipaflokkur1
}
\description{
Skv. \href{http://www.althingi.is/altext/stjt/1997.079.html}{Lögum um veiðar í fiskveiðilandhelgi Íslands}
miðast heimildir fiskiskipa til veiða með botnvörpu, flotvörpu og dragnót innan 
fiskveiðilandhelgi Íslands stærðir skipa og aflvísa þeirra.
}
\details{
1. flokkur: Fiskiskip 42 metrar og lengri. Enn fremur öll fiskiskip með aflvísa 2.500 eða hærri.
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

