% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_log.R
\docType{data}
\name{skipaflokkur2}
\alias{skipaflokkur2}
\title{Heimildir fiskiskipa til veiða í 2. flokki}
\format{\code{SpatialPolygons}}
\usage{
skipaflokkur2
}
\description{
Skv. \href{http://www.althingi.is/altext/stjt/1997.079.html}{Lögum um veiðar í fiskveiðilandhelgi Íslands}
miðast heimildir fiskiskipa til veiða með botnvörpu, flotvörpu og dragnót innan 
fiskveiðilandhelgi Íslands stærðir skipa og aflvísa þeirra.
}
\details{
2. flokkur: Fiskiskip lengri en 29 metrar en styttri en 42 metrar með aflvísa 
lægri en 2.500. Enn fremur fiskiskip styttri en 29 metrar en með aflvísa 
1.600 og hærri. Í þennan flokk falla einnig fiskiskip 39 metrar og styttri 
sem togveiðiheimildir höfðu eftir þeirri viðmiðun skv. 
\href{http://www.althingi.is/lagas/nuna/1976081.html}{3. gr. laga nr. 81 31. maí 1976},
um veiðar í fiskveiðilandhelgi Íslands, enda verði ekki um 
aukningu á aflvísum þeirra að ræða eftir 1. júní 1997.
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

