% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_log.R
\docType{data}
\name{skipaflokkur3}
\alias{skipaflokkur3}
\title{Heimildir fiskiskipa til veiða í 2. flokki}
\format{\code{SpatialPolygons}}
\usage{
skipaflokkur3
}
\description{
Skv. \href{http://www.althingi.is/altext/stjt/1997.079.html}{Lögum um veiðar í fiskveiðilandhelgi Íslands}
miðast heimildir fiskiskipa til veiða með botnvörpu, flotvörpu og dragnót innan
fiskveiðilandhelgi Íslands stærðir skipa og aflvísa þeirra.
}
\details{
3. flokkur: Fiskiskip styttri en 29 metrar, enda séu þau með lægri aflvísa en
1.600. Enn fremur fiskiskip 26 metrar og styttri sem togveiðiheimildir höfðu
eftir þeirri viðmiðun skv.
\href{http://www.althingi.is/lagas/nuna/1976081.html}{3. gr. laga nr. 81 31. maí 1976},
um veiðar í fiskveiðilandhelgi Íslands, enda verði ekki um aukningu á aflvísum
þeirra að ræða eftir 1. júní 1997.
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

