% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{sp_to_geo}
\alias{sp_to_geo}
\title{Convert spatial object to data.frame that can be plotted in the geo-package}
\usage{
sp_to_geo(x, hole)
}
\arguments{
\item{x}{Typically a spatial object}

\item{hole}{A boolean (TRUE or FALSE). If not specified (default) then
everything is returned.}
}
\value{
A \code{data.frame} containg columns lat and lon
}
\description{
Extracts coordinaes from each spatial object into a \code{data.frame} and
separates each with a row of NA's ("lifts the pen" effect).
}
\details{
Can take some time (function currently based on a for-loop)
}
\examples{
d <- sp_to_geo(landhelgi, hole=FALSE)
require(geo)
geoplot()
geolines(d,col="red")
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}

