% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_misc.R
\docType{data}
\name{stratas.sp}
\alias{stratas.sp}
\title{stratas.sp}
\format{\code{SpatialPolygonsDataFrame}}
\usage{
stratas.sp
}
\description{
STRATAS as spatialPolygonDataFrame
}
\details{
Modified object from \code{/net/hafkaldi/export/u2/reikn/Splus5/SMB/GEOMETRY.NEW/.RData}
}
\author{
Höskuldur Björnsson <hoski@hafro.is>
}
\keyword{datasets}

