% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_misc.R
\docType{data}
\name{stratasnew.sp}
\alias{stratasnew.sp}
\title{stratasnew.sp}
\format{\code{SpatialPolygonsDataFrame}}
\usage{
stratasnew.sp
}
\description{
new STRATAS as spatialPolygonDataFrame
}
\details{
Modified object from \code{/net/hafkaldi/export/u2/reikn/Splus5/HAUSTRALLNewStrata/Stratifiering/.RData}
}
\author{
Höskuldur Björnsson <hoski@hafro.is>
}
\keyword{datasets}

