% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmstools.R
\name{tidy_eflalo}
\alias{tidy_eflalo}
\title{tidy eflalo types of data}
\usage{
tidy_eflalo(d, drop_columns = FALSE, wide = TRUE, overlaps = FALSE)
}
\arguments{
\item{d}{eflalo type of dataframe}

\item{drop_columns}{drop columns used for deriving time}

\item{wide}{TRUE (default) returns orginal wide format}

\item{overlaps}{FALSE (default) trips interval overlap. If set to TRUE
will ...}
}
\value{
dataframe or list
}
\description{
tidy eflalo types of data
}

