% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tile_area}
\alias{tile_area}
\title{Calculate area of a tile given resolution}
\usage{
tile_area(x, y, dx, dy)
}
\arguments{
\item{x}{tile horizontal midpoint}

\item{y}{tile vertical midpoint}

\item{dx}{tile horizontal resulution}

\item{dy}{tile vertical resolution}
}
\value{
a numeric vector of length 1
}
\description{
Calculate area of a tile given resolution
}

