% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_shapefiles.R
\name{trim_shape_df}
\alias{trim_shape_df}
\title{trim_shape_df}
\usage{
trim_shape_df(x, xlim = c(-30, 10), ylim = c(64, 67))
}
\arguments{
\item{x}{A shape data frame}

\item{xlim}{Longitudinal boundaries}

\item{ylim}{Latitudinal boundaries}
}
\description{
Trims shape data.frames and returns a shape data.frame
}

