% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_log.R
\docType{data}
\name{vidmidunarlina}
\alias{vidmidunarlina}
\title{Viðmiðunarlína}
\format{\code{SpatialPolygonsDataFrame}}
\usage{
vidmidunarlina
}
\description{
Línur sem eru lagðar til grundvallar á útreikningi á lögum um veiðar í landhelgi sbr.
 \href{http://www.althingi.is/lagas/nuna/1997079.html}{Lög um veiðar í fiskveiðilandhelgi Íslands}
}
\details{
Athugið að stórstraumsfjöruborð Vestmannaeyja er hér meðtalið
}
\examples{
# Fjórar sjómílur út frá viðmiðunarlínu (þ.m.t. frá Vestmannaeyjum)
sp::plot(expand_sp(vidmidunarlina, miles=4), col=scales::alpha('red',0.2),lwd=0.001,border='red')
sp::plot(iceland, col="grey90", add=TRUE)
sp::plot(vidmidunarlina, border="yellow", lwd=2, add=TRUE)
sp::plot(vidmidunarpunktar, col="red", add=TRUE)
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

