% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_log.R
\docType{data}
\name{vidmidunarpunktar}
\alias{vidmidunarpunktar}
\title{Viðmiðunarpunktar}
\format{\code{SpatialPointsDataFrame}}
\source{
\url{http://www.althingi.is/lagas/nuna/1997079.html}
}
\usage{
vidmidunarpunktar
}
\description{
Hnit sem eru lögð til grundvallar á útreikningi á lögum um veiðar í landhelgi sbr. 
 \href{http://www.althingi.is/lagas/nuna/1997079.html}{Lög um veiðar í fiskveiðilandhelgi Íslands}
}
\examples{
sp::plot(iceland,col="grey90")
sp::plot(vidmidunarpunktar, col="red", add=TRUE)
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

