% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_misc.R
\docType{data}
\name{bormicon}
\alias{bormicon}
\title{bormicon}
\format{\code{SpatialPolygonsDataFrame}}
\usage{
bormicon
}
\description{
Bormicon areas
}
\details{
Modified object from the \code{fjolst}-package
}
\author{
Höskuldur Björnsson <hoski@hafro.is>
}
\keyword{datasets}

