% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_log.R
\docType{data}
\name{eez}
\alias{eez}
\title{Íslenska efnahagslögsagan ('200 mílur')}
\format{An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 3 columns.}
\usage{
eez
}
\description{
Íslenska efnahaglögsagan reiknuð út frá grunnlínum
(\code{\link{grunnlina}}) sbr.
\href{http://www.althingi.is/altext/lagas/141b/1979041.html}{Lög um landhelgi, efnahagslögsögu og landgrunn}
Hnit á miðlínu eru fengin úr milliríkjasamningum milli 
\href{http://www.althingi.is/altext/133/s/0951.html}{Íslands og Færeyjar} og milli
\href{http://www.althingi.is/altext/122/s/1047.html}{Íslands og Grænlands/Noregs}.
}
\examples{
sp::plot(eez,col=scales::alpha('red',0.2),border="red")
sp::plot(landhelgi,col=scales::alpha('red',0.2),border="red",add=TRUE)
sp::plot(iceland,col="grey90",add=TRUE)
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

