% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_log.R
\docType{data}
\name{eez_sldf}
\alias{eez_sldf}
\title{Íslenska efnahagslögsagan ('200 mílur') - Línur}
\format{An object of class \code{SpatialLinesDataFrame} with 1 rows and 3 columns.}
\usage{
eez_sldf
}
\description{
Íslenska efnahaglögsagan reiknuð út frá grunnlínum
(\code{\link{grunnlina}}) sbr.
\href{http://www.althingi.is/altext/lagas/141b/1979041.html}{Lög um landhelgi, efnahagslögsögu og landgrunn}
Hnit á miðlínu eru fengin úr milliríkjasamningum milli 
\href{http://www.althingi.is/altext/133/s/0951.html}{Íslands og Færeyjar} og milli
\href{http://www.althingi.is/altext/122/s/1047.html}{Íslands og Grænlands/Noregs}.
}
\examples{
sp::plot(eez_sldf, col="blue")
sp::plot(landhelgi_sldf,col="red",add=TRUE)
sp::plot(iceland,col="grey90",add=TRUE)
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

