% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_sjokort.R
\name{get_sjokort}
\alias{get_sjokort}
\title{Get Sjókort}
\usage{
get_sjokort(bbox = c(left = -24, bottom = 64, right = -22, top = 65),
  zoom = 8, maptype = c("sjm"), crop = TRUE, messaging = FALSE,
  urlonly = FALSE, color = c("color"), force = FALSE, where = tempdir(),
  ...)
}
\arguments{
\item{bbox}{a bounding box in the format c(lowerleftlon, lowerleftlat,
upperrightlon, upperrightlat).}

\item{zoom}{a zoom level}

\item{maptype}{terrain, terrain-background, terrain-labels, terrain-lines,
toner, toner-2010, toner-2011, toner-background, toner-hybrid,
toner-labels, toner-lines, toner-lite, or watercolor.}

\item{crop}{crop raw map tiles to specified bounding box}

\item{messaging}{turn messaging on/off}

\item{urlonly}{return url only}

\item{color}{color or black-and-white}

\item{force}{if the map is on file, should a new map be looked up?}

\item{where}{where should the file drawer be located (without terminating "/")}

\item{...}{...}
}
\value{
a ggmap object (a classed raster object with a bounding box attribute)
}
\description{
\code{get_stamenmap} accesses a tile server for Stamen Maps and
downloads/stitches map tiles/formats a map image. Note that Stamen maps don't
cover the entire world, e.g. \url{http://tile.stamen.com/terrain/#4/30.28/-87.21}
}
\seealso{
\url{http://maps.stamen.com/#watercolor}, \code{\link{ggmap}}
}

