% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_reglugerdir.R
\docType{data}
\name{reglugerdir}
\alias{reglugerdir}
\title{Reglugerðir um stjórn fiskveiða}
\format{\preformatted{Formal class 'SpatialPolygonsDataFrame' [package "sp"] with 5 slots
  ..@ data       :'data.frame':	46 obs. of  22 variables:
  .. ..$ id        : num [1:46] 1 1 1 1 2 3 1 2 1 2 ...
  .. ..$ type      : Factor w/ 5 levels "permanent","spawning",..: 1 1 1 1 1 1 2 2 3 3 ...
  .. ..$ year      : int [1:46] 2002 2003 2003 2003 2003 2003 2004 2004 2005 2005 ...
  .. ..$ nr        : int [1:46] 162 68 230 311 311 311 202 202 1140 1140 ...
  .. ..$ grein     : int [1:46] 1 1 1 1 2 3 1 1 1 1 ...
  .. ..$ tolulidur : chr [1:46] NA NA NA NA ...
  .. ..$ undirlidur: chr [1:46] NA NA NA NA ...
  .. ..$ heiti     : chr [1:46] "Malarrif" "Sléttugrunn" "Kötlugrunn" "Síðugrunn" ...
  .. ..$ dags_fra  : Date[1:46], format: "2002-03-03" "2003-02-07" ...
  .. ..$ dags_til  : Date[1:46], format: NA NA ...
  .. ..$ v1        : int [1:46] 0 1 1 1 1 1 1 1 1 1 ...
  .. ..$ v2        : int [1:46] 0 0 0 0 0 0 0 0 1 1 ...
  .. ..$ v3        : int [1:46] 0 0 0 0 0 0 0 0 1 1 ...
  .. ..$ v5        : int [1:46] 0 0 0 0 0 0 0 0 1 1 ...
  .. ..$ v6        : int [1:46] 1 1 0 0 0 0 1 1 1 1 ...
  .. ..$ v7        : int [1:46] 0 0 0 0 0 0 0 0 0 0 ...
  .. ..$ v8        : int [1:46] 0 0 0 0 0 0 0 0 0 0 ...
  .. ..$ v9        : int [1:46] 0 0 0 0 0 0 0 0 1 1 ...
  .. ..$ v14       : int [1:46] 0 0 0 0 0 0 0 0 1 1 ...
  .. ..$ title     : Factor w/ 21 levels "Um bann við veiðum með fiskibotnvörpu suðvestur af Malarrifi.",..: 1 2 3 4 4 4 5 5 6 6 ...
  .. ..$ ref       : Factor w/ 21 levels "http://www.reglugerd.is/interpro/dkm/WebGuard.nsf/aa0d47377abc977400256a090053ff91/b075510aa14af70b00256b6f003ec9f5?OpenDocumen"| __truncated__,..: 1 2 3 4 4 4 5 5 6 6 ...
  .. ..$ area      : num [1:46] 10.9 217.3 532.1 205 371.3 ...
  ..@ polygons   :List of 46
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -24.5 64.5
  .. .. .. .. .. .. ..@ area   : num 0.00203
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:5, 1:2] -24.6 -24.4 -24.4 -24.5 -24.6 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:5] "1" "2" "3" "4" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -24.5 64.5
  .. .. .. ..@ ID       : chr "1"
  .. .. .. ..@ area     : num 0.00203
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -17.3 67
  .. .. .. .. .. .. ..@ area   : num 0.0444
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:5, 1:2] -17.4 -17.3 -17.1 -17.3 -17.4 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:5] "5" "6" "7" "8" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -17.3 67
  .. .. .. ..@ ID       : chr "2"
  .. .. .. ..@ area     : num 0.0444
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -18.4 63.3
  .. .. .. .. .. .. ..@ area   : num 0.095
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:5, 1:2] -18.8 -18.1 -18.1 -18.7 -18.8 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:5] "9" "10" "11" "12" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -18.4 63.3
  .. .. .. ..@ ID       : chr "3"
  .. .. .. ..@ area     : num 0.095
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -17.5 63.5
  .. .. .. .. .. .. ..@ area   : num 0.0369
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:5, 1:2] -17.3 -17.3 -17.6 -17.6 -17.3 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:5] "13" "14" "15" "16" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -17.5 63.5
  .. .. .. ..@ ID       : chr "4"
  .. .. .. ..@ area     : num 0.0369
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -16.6 63.7
  .. .. .. .. .. .. ..@ area   : num 0.0672
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:6, 1:2] -16.6 -16.3 -16.6 -16.9 -16.8 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:6] "17" "18" "19" "20" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -16.6 63.7
  .. .. .. ..@ ID       : chr "5"
  .. .. .. ..@ area     : num 0.0672
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -15.6 63.9
  .. .. .. .. .. .. ..@ area   : num 0.07
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:5, 1:2] -15.4 -15.4 -15.8 -15.9 -15.4 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:5] "22" "23" "24" "25" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -15.6 63.9
  .. .. .. ..@ ID       : chr "6"
  .. .. .. ..@ area     : num 0.07
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -20.2 62.9
  .. .. .. .. .. .. ..@ area   : num 0.0709
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:6, 1:2] -20.4 -20 -20 -20.4 -20.4 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:6] "26" "27" "28" "29" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -20.2 62.9
  .. .. .. ..@ ID       : chr "7"
  .. .. .. ..@ area     : num 0.0709
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -27.5 61
  .. .. .. .. .. .. ..@ area   : num 0.00652
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:4, 1:2] -27.7 -27.5 -27.5 -27.7 61 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:4] "31" "32" "33" "311"
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -27.5 61
  .. .. .. ..@ ID       : chr "8"
  .. .. .. ..@ area     : num 0.00652
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -14.8 63.6
  .. .. .. .. .. .. ..@ area   : num 0.00143
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:7, 1:2] -14.8 -14.8 -14.8 -14.8 -14.7 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:7] "34" "35" "36" "37" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -14.8 63.6
  .. .. .. ..@ ID       : chr "9"
  .. .. .. ..@ area     : num 0.00143
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -14.7 63.6
  .. .. .. .. .. .. ..@ area   : num 0.00582
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:9, 1:2] -14.6 -14.7 -14.8 -14.8 -14.8 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:9] "40" "41" "42" "43" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -14.7 63.6
  .. .. .. ..@ ID       : chr "10"
  .. .. .. ..@ area     : num 0.00582
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -17.9 63.2
  .. .. .. .. .. .. ..@ area   : num 0.00131
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:8, 1:2] -17.9 -17.9 -17.9 -17.8 -17.8 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:8] "48" "49" "50" "51" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -17.9 63.2
  .. .. .. ..@ ID       : chr "11"
  .. .. .. ..@ area     : num 0.00131
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -17.8 63.2
  .. .. .. .. .. .. ..@ area   : num 0.00397
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:8, 1:2] -18 -18 -17.9 -17.6 -17.6 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:8] "55" "56" "57" "58" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -17.8 63.2
  .. .. .. ..@ ID       : chr "12"
  .. .. .. ..@ area     : num 0.00397
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -18.8 63.2
  .. .. .. .. .. .. ..@ area   : num 0.00169
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:8, 1:2] -18.7 -18.7 -18.7 -18.8 -18.9 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:8] "62" "63" "64" "65" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -18.8 63.2
  .. .. .. ..@ ID       : chr "13"
  .. .. .. ..@ area     : num 0.00169
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -20.6 63.3
  .. .. .. .. .. .. ..@ area   : num 0.00581
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:6, 1:2] -20.6 -20.7 -20.7 -20.6 -20.5 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:6] "69" "70" "71" "72" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -20.6 63.3
  .. .. .. ..@ ID       : chr "14"
  .. .. .. ..@ area     : num 0.00581
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -20.4 63.5
  .. .. .. .. .. .. ..@ area   : num 0.0224
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:6, 1:2] -20.6 -20.3 -20.3 -20.3 -20.6 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:6] "74" "75" "76" "77" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -20.4 63.5
  .. .. .. ..@ ID       : chr "15"
  .. .. .. ..@ area     : num 0.0224
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -24.1 65.3
  .. .. .. .. .. .. ..@ area   : num 0.0958
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:404, 1:2] -24.2 -24.2 -24.2 -24.2 -24.2 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : NULL
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "x" "y"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -24.1 65.3
  .. .. .. ..@ ID       : chr "16"
  .. .. .. ..@ area     : num 0.0958
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -24.5 65.4
  .. .. .. .. .. .. ..@ area   : num 0.0565
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:334, 1:2] -24.5 -24.5 -24.5 -24.5 -24.5 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : NULL
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "x" "y"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -24.5 65.4
  .. .. .. ..@ ID       : chr "17"
  .. .. .. ..@ area     : num 0.0565
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 2
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -26.5 64.5
  .. .. .. .. .. .. ..@ area   : num 1.06
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:21, 1:2] -26.6 -26.8 -26.8 -26.7 -26.1 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:21] "1" "2" "3" "4" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "long" "lat"
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -24.6 63.1
  .. .. .. .. .. .. ..@ area   : num 0.0495
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:10, 1:2] -24.7 -24.4 -24.4 -24.4 -24.5 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:10] "23" "24" "25" "26" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "long" "lat"
  .. .. .. ..@ plotOrder: int [1:2] 1 2
  .. .. .. ..@ labpt    : num [1:2] -26.5 64.5
  .. .. .. ..@ ID       : chr "18"
  .. .. .. ..@ area     : num 1.11
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -25.7 64
  .. .. .. .. .. .. ..@ area   : num 0.22
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:7, 1:2] -25.6 -25.1 -25.7 -26.1 -26.2 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:7] "111" "112" "113" "114" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -25.7 64
  .. .. .. ..@ ID       : chr "19"
  .. .. .. ..@ area     : num 0.22
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -25.2 63.3
  .. .. .. .. .. .. ..@ area   : num 0.119
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:6, 1:2] -25.4 -24.7 -24.9 -25.5 -25.7 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:6] "117" "118" "119" "120" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -25.2 63.3
  .. .. .. ..@ ID       : chr "20"
  .. .. .. ..@ area     : num 0.119
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -22.4 63.1
  .. .. .. .. .. .. ..@ area   : num 0.0402
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:10, 1:2] -22 -21.9 -22.1 -22.4 -22.5 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:10] "122" "123" "124" "125" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -22.4 63.1
  .. .. .. ..@ ID       : chr "21"
  .. .. .. ..@ area     : num 0.0402
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -15.8 64
  .. .. .. .. .. .. ..@ area   : num 0.116
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:39, 1:2] -15.7 -15.7 -15.7 -15.5 -15.8 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : NULL
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "x" "y"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -15.8 64
  .. .. .. ..@ ID       : chr "22"
  .. .. .. ..@ area     : num 0.116
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -22.4 66.8
  .. .. .. .. .. .. ..@ area   : num 0.42
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:271, 1:2] -23.4 -23.4 -23.7 -22 -22 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : NULL
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "x" "y"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -22.4 66.8
  .. .. .. ..@ ID       : chr "23"
  .. .. .. ..@ area     : num 0.42
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -21.3 66.7
  .. .. .. .. .. .. ..@ area   : num 0.581
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:8, 1:2] -21.8 -21.8 -20.8 -20.7 -21.2 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:8] "142" "143" "144" "145" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -21.3 66.7
  .. .. .. ..@ ID       : chr "24"
  .. .. .. ..@ area     : num 0.581
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -19.8 66.7
  .. .. .. .. .. .. ..@ area   : num 0.15
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:6, 1:2] -20.3 -20.3 -19.9 -19.4 -19.4 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:6] "149" "150" "151" "152" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -19.8 66.7
  .. .. .. ..@ ID       : chr "25"
  .. .. .. ..@ area     : num 0.15
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -19.2 66.4
  .. .. .. .. .. .. ..@ area   : num 0.0542
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:5, 1:2] -19.5 -19.5 -19 -18.9 -19.5 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:5] "154" "155" "156" "157" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -19.2 66.4
  .. .. .. ..@ ID       : chr "26"
  .. .. .. ..@ area     : num 0.0542
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -15 66.7
  .. .. .. .. .. .. ..@ area   : num 0.259
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:489, 1:2] -15.9 -15.9 -15.8 -15.8 -15.8 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : NULL
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "x" "y"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -15 66.7
  .. .. .. ..@ ID       : chr "27"
  .. .. .. ..@ area     : num 0.259
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -13.7 66.7
  .. .. .. .. .. .. ..@ area   : num 0.199
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:11, 1:2] -14 -14.2 -14 -13.8 -13.5 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:11] "158" "159" "160" "161" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -13.7 66.7
  .. .. .. ..@ ID       : chr "28"
  .. .. .. ..@ area     : num 0.199
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -13.2 66.2
  .. .. .. .. .. .. ..@ area   : num 0.319
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:7, 1:2] -13.8 -12.8 -12.7 -12.6 -12.9 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:7] "168" "169" "170" "171" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -13.2 66.2
  .. .. .. ..@ ID       : chr "29"
  .. .. .. ..@ area     : num 0.319
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -21.3 66.8
  .. .. .. .. .. .. ..@ area   : num 0.487
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:9, 1:2] -21.8 -21.8 -20.8 -20.7 -21.2 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:9] "174" "175" "176" "177" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -21.3 66.8
  .. .. .. ..@ ID       : chr "30"
  .. .. .. ..@ area     : num 0.487
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -21.4 66.8
  .. .. .. .. .. .. ..@ area   : num 0.386
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:10, 1:2] -21.8 -21.8 -20.8 -20.7 -20.9 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:10] "182" "183" "184" "185" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -21.4 66.8
  .. .. .. ..@ ID       : chr "31"
  .. .. .. ..@ area     : num 0.386
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -20.8 63.6
  .. .. .. .. .. .. ..@ area   : num 0.065
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:5, 1:2] -20.6 -20.6 -21 -21 -20.6 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:5] "191" "192" "193" "194" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -20.8 63.6
  .. .. .. ..@ ID       : chr "32"
  .. .. .. ..@ area     : num 0.065
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -21.3 66.9
  .. .. .. .. .. .. ..@ area   : num 0.299
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:8, 1:2] -21.8 -21.8 -20.8 -20.7 -20.9 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:8] "195" "196" "197" "198" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -21.3 66.9
  .. .. .. ..@ ID       : chr "33"
  .. .. .. ..@ area     : num 0.299
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -19.3 66.2
  .. .. .. .. .. .. ..@ area   : num 0.0521
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:6, 1:2] -19.1 -19.6 -19.4 -19.2 -19 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:6] "202" "203" "204" "205" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -19.3 66.2
  .. .. .. ..@ ID       : chr "34"
  .. .. .. ..@ area     : num 0.0521
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -23.3 65.1
  .. .. .. .. .. .. ..@ area   : num 0.0504
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:8, 1:2] -23.4 -23.2 -23.2 -23.2 -23.5 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:8] "207" "208" "209" "210" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -23.3 65.1
  .. .. .. ..@ ID       : chr "35"
  .. .. .. ..@ area     : num 0.0504
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -16.3 63.4
  .. .. .. .. .. .. ..@ area   : num 0.0447
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:5, 1:2] -16.7 -16 -16 -16.6 -16.7 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:5] "214" "215" "216" "217" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -16.3 63.4
  .. .. .. ..@ ID       : chr "36"
  .. .. .. ..@ area     : num 0.0447
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -15.6 63.7
  .. .. .. .. .. .. ..@ area   : num 0.0989
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:6, 1:2] -15.4 -15.4 -15.6 -15.8 -15.8 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:6] "218" "219" "220" "221" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -15.6 63.7
  .. .. .. ..@ ID       : chr "37"
  .. .. .. ..@ area     : num 0.0989
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -14.4 63.9
  .. .. .. .. .. .. ..@ area   : num 0.368
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:10, 1:2] -14.2 -13.8 -14.1 -14.5 -14.8 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:10] "223" "224" "225" "226" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -14.4 63.9
  .. .. .. ..@ ID       : chr "38"
  .. .. .. ..@ area     : num 0.368
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -20.4 63.4
  .. .. .. .. .. .. ..@ area   : num 0.00667
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:5, 1:2] -20.4 -20.5 -20.5 -20.4 -20.4 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:5] "232" "233" "234" "235" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -20.4 63.4
  .. .. .. ..@ ID       : chr "39"
  .. .. .. ..@ area     : num 0.00667
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -16.9 63.3
  .. .. .. .. .. .. ..@ area   : num 0.0116
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:11, 1:2] -17.2 -17.2 -17 -16.9 -16.8 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:11] "236" "237" "238" "239" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -16.9 63.3
  .. .. .. ..@ ID       : chr "40"
  .. .. .. ..@ area     : num 0.0116
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -14 63.9
  .. .. .. .. .. .. ..@ area   : num 0.014
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:7, 1:2] -14 -14 -13.9 -14.1 -14.1 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:7] "246" "247" "248" "249" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -14 63.9
  .. .. .. ..@ ID       : chr "41"
  .. .. .. ..@ area     : num 0.014
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -13.8 63.8
  .. .. .. .. .. .. ..@ area   : num 0.0143
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:15, 1:2] -14.1 -14 -13.9 -13.8 -13.8 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:15] "252" "253" "254" "255" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -13.8 63.8
  .. .. .. ..@ ID       : chr "42"
  .. .. .. ..@ area     : num 0.0143
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -13.5 64
  .. .. .. .. .. .. ..@ area   : num 0.00317
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:6, 1:2] -13.6 -13.5 -13.4 -13.5 -13.6 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:6] "266" "267" "268" "269" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -13.5 64
  .. .. .. ..@ ID       : chr "43"
  .. .. .. ..@ area     : num 0.00317
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -13.2 63.5
  .. .. .. .. .. .. ..@ area   : num 0.0296
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:8, 1:2] -13.3 -13.3 -13.3 -13.3 -13.3 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:8] "271" "272" "273" "274" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -13.2 63.5
  .. .. .. ..@ ID       : chr "44"
  .. .. .. ..@ area     : num 0.0296
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -11.3 65
  .. .. .. .. .. .. ..@ area   : num 0.153
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:8, 1:2] -11.5 -11.1 -11.1 -11.3 -11.6 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:8] "278" "279" "280" "281" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -11.3 65
  .. .. .. ..@ ID       : chr "45"
  .. .. .. ..@ area     : num 0.153
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
  .. .. .. ..@ Polygons :List of 1
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
  .. .. .. .. .. .. ..@ labpt  : num [1:2] -27.6 61
  .. .. .. .. .. .. ..@ area   : num 0.0145
  .. .. .. .. .. .. ..@ hole   : logi FALSE
  .. .. .. .. .. .. ..@ ringDir: int 1
  .. .. .. .. .. .. ..@ coords : num [1:5, 1:2] -27.5 -27.5 -27.7 -27.7 -27.5 ...
  .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. .. .. .. .. ..$ : chr [1:5] "285" "286" "287" "288" ...
  .. .. .. .. .. .. .. .. ..$ : chr [1:2] "lon" "lat"
  .. .. .. ..@ plotOrder: int 1
  .. .. .. ..@ labpt    : num [1:2] -27.6 61
  .. .. .. ..@ ID       : chr "46"
  .. .. .. ..@ area     : num 0.0145
  ..@ plotOrder  : int [1:46] 18 24 30 23 31 38 29 33 27 19 ...
  ..@ bbox       : num [1:2, 1:2] -27.7 60.9 -11.1 67.1
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:2] "x" "y"
  .. .. ..$ : chr [1:2] "min" "max"
  ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
  .. .. ..@ projargs: chr "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"
}}
\usage{
reglugerdir
}
\description{
Incomplete
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

