% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{scale_latitude}
\alias{scale_latitude}
\title{Title}
\usage{
scale_latitude(ymin = -90, ymax = 90, step = 0.5, ...)
}
\arguments{
\item{ymin}{XXX}

\item{ymax}{XXX}

\item{step}{XXX}

\item{...}{XXX}
}
\description{
Title
}

