% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{scale_latitude}
\alias{scale_latitude}
\title{Title}
\usage{
scale_latitude(from = -90, to = 90, by = 0.5, ...)
}
\arguments{
\item{from}{XXX}

\item{to}{XXX}

\item{by}{XXX}

\item{...}{XXX}
}
\description{
Title
}

