% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{scale_longitude}
\alias{scale_longitude}
\title{Title}
\usage{
scale_longitude(from = -180, to = 180, step = 1, ...)
}
\arguments{
\item{from}{XXX}

\item{to}{XXX}

\item{step}{XXX}

\item{...}{XXX}
}
\description{
Title
}

